# openapi.yaml
...
openapi: 3.0.0
paths:
...
  /v1/payments:
    post:
      description: >-
        Creates a Payment and hosted checkout session. Requires Idempotency-Key. Returns the same payment on replay with
        the same key and body.
      operationId: PaymentsController_create
      parameters:
        - name: X-Request-Id
          in: header
          description: Optional client correlation id (echoed on the response)
          required: false
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          description: Unique key (1–255 chars) to guarantee at-most-once creation
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentDto'
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponseDto'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '401':
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '409':
          description: Idempotency key reused with a different body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
      security:
        - ApiKey: []
      summary: Create a payment
      tags:
        - Payments
    get:
      description: Cursor-paginated list of payments for the authenticated merchant/environment.
      operationId: PaymentsController_findAll
      parameters:
        - name: limit
          required: false
          in: query
          description: Number of results to return (default 25, max 100)
          schema:
            minimum: 1
            maximum: 100
            default: 25
            type: number
        - name: cursor
          required: false
          in: query
          description: Opaque cursor from a previous list response next_cursor field
          schema:
            type: string
      responses:
        '200':
          description: Paginated list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentListResponseDto'
        '401':
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
      security:
        - ApiKey: []
      summary: List payments
      tags:
        - Payments
  /v1/payments/{id}:
    get:
      operationId: PaymentsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Payment publicId (pay_...)
          schema:
            type: string
      responses:
        '200':
          description: Payment object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponseDto'
        '401':
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
      security:
        - ApiKey: []
      summary: Retrieve a payment
      tags:
        - Payments
  /v1/payments/{id}/cancel:
    post:
      description: Cancels a payment in created or pending status.
      operationId: PaymentsController_cancel
      parameters:
        - name: id
          required: true
          in: path
          description: Payment publicId (pay_...)
          schema:
            type: string
      responses:
        '200':
          description: Canceled payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponseDto'
        '401':
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '409':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
      security:
        - ApiKey: []
      summary: Cancel a payment
      tags:
        - Payments
  /v1/payments/{id}/events:
    get:
      operationId: PaymentsController_findEvents
      parameters:
        - name: id
          required: true
          in: path
          description: Payment publicId (pay_...)
          schema:
            type: string
      responses:
        '200':
          description: Immutable payment event log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentEventListResponseDto'
        '401':
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
      security:
        - ApiKey: []
      summary: List events for a payment
      tags:
        - Payments
info:
  title: Hypertron Payments API
  description: >-
    Production-grade stablecoin payment gateway API (HyperTone). Public merchant API under `/v1/*` (API key auth).
    Dashboard control-plane under `/api/developer/*` (Freighter session cookie). Contract reference:
    Payments_API_v1_Schema.md.
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3000
    description: Local
  - url: https://api.hypertron.xyz
    description: Production
components:
  securitySchemes:
    ApiKey:
      scheme: bearer
      bearerFormat: sk_test_... / sk_live_...
      type: http
      description: Secret API key for /v1 routes
    SessionCookie:
      type: apiKey
      in: cookie
      name: ht_dashboard
      description: Freighter session cookie from hypertron-core-backend (HMAC AUTH_SECRET)
    IdempotencyKey:
      type: apiKey
      in: header
      name: Idempotency-Key
      description: Required for POST /v1/payments
    RequestId:
      type: apiKey
      in: header
      name: X-Request-Id
      description: Optional client correlation id
  schemas:
    HypertronErrorBodyDto:
      type: object
      properties:
        type:
          type: string
          example: invalid_request_error
          enum:
            - invalid_request_error
            - authentication_error
            - permission_error
            - resource_missing
            - idempotency_error
            - invalid_state_transition
            - unprocessable_entity
            - rate_limit_error
            - api_error
            - service_unavailable
        code:
          type: string
          example: invalid_amount
        message:
          type: string
          example: amount must be a positive decimal string
        param:
          type: string
          example: amount
          description: Present only for field-level validation errors
        request_id:
          type: string
          example: req_01J...
      required:
        - type
        - code
        - message
        - request_id
    HypertronErrorResponseDto:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/HypertronErrorBodyDto'
      required:
        - error
...
    CreatePaymentDto:
      type: object
      properties:
        amount:
          type: string
          description: Payment amount as a positive decimal string (no scientific notation)
          example: '10.50'
        currency:
          type: string
          description: Currency code
          enum:
            - USDC
            - EURC
            - XLM
          example: USDC
        description:
          type: string
          description: Human-readable description shown on the checkout page
          example: 'Order #1234'
          maxLength: 500
        customer_email:
          type: string
          description: Customer email — used to upsert the merchant-scoped customer record
          example: alice@example.com
        customer_name:
          type: string
          description: Customer display name
          example: Alice Smith
        metadata:
          type: object
          description: Key-value metadata (max 50 pairs, values max 500 chars)
          example:
            order_id: ord_123
          additionalProperties:
            type: string
      required:
        - amount
        - currency
    PaymentResponseDto:
      type: object
      properties:
        id:
          type: string
          example: pay_01JABC123
        object:
          type: string
          example: payment
          enum:
            - payment
        business_id:
          type: string
          example: clxyz...
        environment:
          type: string
          enum:
            - test
            - live
          example: test
        status:
          type: string
          enum:
            - created
            - pending
            - confirmed
            - completed
            - failed
            - expired
            - canceled
          example: pending
        amount:
          type: string
          example: '100.00'
          description: Decimal string amount
        currency:
          type: string
          enum:
            - USDC
            - EURC
            - XLM
          example: USDC
        description:
          type: object
          example: Order ORD123
          nullable: true
        customer_id:
          type: object
          example: cus_01JABC456
          nullable: true
        metadata:
          type: object
          example:
            order_id: ORD123
          additionalProperties: true
        checkout_url:
          type: string
          example: https://pay.hypertron.xyz/pay/...
        link_memo:
          type: string
          example: hpl_abc123
        destination_address:
          type: string
          example: G...
        payer_address:
          type: object
          example: G...
          nullable: true
        transaction_hash:
          type: object
          example: a1b2c3...
          nullable: true
        failure_code:
          type: object
          example: wrong_asset
          nullable: true
        failure_message:
          type: object
          example: Asset code mismatch
          nullable: true
        expires_at:
          type: object
          example: '2026-08-03T13:30:00.000Z'
          nullable: true
        paid_at:
          type: object
          example: '2026-08-03T12:31:00.000Z'
          nullable: true
        completed_at:
          type: object
          example: '2026-08-03T12:31:10.000Z'
          nullable: true
        canceled_at:
          type: object
          example: null
          nullable: true
        created_at:
          type: string
          example: '2026-08-03T12:30:00.000Z'
        updated_at:
          type: string
          example: '2026-08-03T12:30:00.000Z'
      required:
        - id
        - object
        - business_id
        - environment
        - status
        - amount
        - currency
        - metadata
        - checkout_url
        - link_memo
        - destination_address
        - created_at
        - updated_at
    PaymentListResponseDto:
      type: object
      properties:
        object:
          type: string
          example: list
          enum:
            - list
        data:
          type: array
          items:
            $ref: '#/components/schemas/PaymentResponseDto'
        has_more:
          type: boolean
          example: false
        next_cursor:
          type: object
          example: null
          nullable: true
      required:
        - object
        - data
        - has_more
    PaymentEventResponseDto:
      type: object
      properties:
        id:
          type: string
          example: evt_01JABC789
        object:
          type: string
          example: payment_event
          enum:
            - payment_event
        payment_id:
          type: string
          description: Internal payment ObjectId (opaque to merchants in v1 list)
          example: 665f...
        type:
          type: string
          example: payment.completed
          enum:
            - payment.created
            - payment.pending
            - payment.confirmed
            - payment.completed
            - payment.failed
            - payment.expired
            - payment.canceled
        data:
          type: object
          description: Immutable payment snapshot at event emission time
          additionalProperties: true
        created_at:
          type: string
          example: '2026-08-03T12:31:10.000Z'
      required:
        - id
        - object
        - payment_id
        - type
        - data
        - created_at
    PaymentEventListResponseDto:
      type: object
      properties:
        object:
          type: string
          example: list
          enum:
            - list
        data:
          type: array
          items:
            $ref: '#/components/schemas/PaymentEventResponseDto'
      required:
        - object
        - data